+++ client/api.go
@@ -112,6 +112,7 @@ import (
 	"github.com/stripe/stripe-go/v83/subscription"
 	"github.com/stripe/stripe-go/v83/subscriptionitem"
 	"github.com/stripe/stripe-go/v83/subscriptionschedule"
+	taxassociation "github.com/stripe/stripe-go/v83/tax/association"
 	taxcalculation "github.com/stripe/stripe-go/v83/tax/calculation"
 	taxregistration "github.com/stripe/stripe-go/v83/tax/registration"
 	taxsettings "github.com/stripe/stripe-go/v83/tax/settings"
@@ -122,6 +123,7 @@ import (
 	terminalconfiguration "github.com/stripe/stripe-go/v83/terminal/configuration"
 	terminalconnectiontoken "github.com/stripe/stripe-go/v83/terminal/connectiontoken"
 	terminallocation "github.com/stripe/stripe-go/v83/terminal/location"
+	terminalonboardinglink "github.com/stripe/stripe-go/v83/terminal/onboardinglink"
 	terminalreader "github.com/stripe/stripe-go/v83/terminal/reader"
 	testhelpersconfirmationtoken "github.com/stripe/stripe-go/v83/testhelpers/confirmationtoken"
 	testhelperscustomer "github.com/stripe/stripe-go/v83/testhelpers/customer"
@@ -364,6 +366,8 @@ type API struct {
 	Subscriptions *subscription.Client
 	// SubscriptionSchedules is the client used to invoke /v1/subscription_schedules APIs.
 	SubscriptionSchedules *subscriptionschedule.Client
+	// TaxAssociations is the client used to invoke association related APIs.
+	TaxAssociations *taxassociation.Client
 	// TaxCalculations is the client used to invoke /v1/tax/calculations APIs.
 	TaxCalculations *taxcalculation.Client
 	// TaxCodes is the client used to invoke /v1/tax_codes APIs.
@@ -384,6 +388,8 @@ type API struct {
 	TerminalConnectionTokens *terminalconnectiontoken.Client
 	// TerminalLocations is the client used to invoke /v1/terminal/locations APIs.
 	TerminalLocations *terminallocation.Client
+	// TerminalOnboardingLinks is the client used to invoke /v1/terminal/onboarding_links APIs.
+	TerminalOnboardingLinks *terminalonboardinglink.Client
 	// TerminalReaders is the client used to invoke /v1/terminal/readers APIs.
 	TerminalReaders *terminalreader.Client
 	// TestHelpersConfirmationTokens is the client used to invoke /v1/confirmation_tokens APIs.
@@ -571,6 +577,7 @@ func (a *API) Init(key string, backends *stripe.Backends) {
 	a.SubscriptionItems = &subscriptionitem.Client{B: backends.API, Key: key}
 	a.Subscriptions = &subscription.Client{B: backends.API, Key: key}
 	a.SubscriptionSchedules = &subscriptionschedule.Client{B: backends.API, Key: key}
+	a.TaxAssociations = &taxassociation.Client{B: backends.API, Key: key}
 	a.TaxCalculations = &taxcalculation.Client{B: backends.API, Key: key}
 	a.TaxCodes = &taxcode.Client{B: backends.API, Key: key}
 	a.TaxIDs = &taxid.Client{B: backends.API, Key: key}
@@ -581,6 +588,7 @@ func (a *API) Init(key string, backends *stripe.Backends) {
 	a.TerminalConfigurations = &terminalconfiguration.Client{B: backends.API, Key: key}
 	a.TerminalConnectionTokens = &terminalconnectiontoken.Client{B: backends.API, Key: key}
 	a.TerminalLocations = &terminallocation.Client{B: backends.API, Key: key}
+	a.TerminalOnboardingLinks = &terminalonboardinglink.Client{B: backends.API, Key: key}
 	a.TerminalReaders = &terminalreader.Client{B: backends.API, Key: key}
 	a.TestHelpersConfirmationTokens = &testhelpersconfirmationtoken.Client{B: backends.API, Key: key}
 	a.TestHelpersCustomers = &testhelperscustomer.Client{B: backends.API, Key: key}